move spinner, add story, change candidate weight
