+++ frontend/talentsearch/src/js/components/Spinner/Spinner.stories.tsx
@@ -1,0 +1,8 @@
+import type { Meta, Story } from "@storybook/react";
+
+import Spinner from "./Spinner";
+
+export default {
+  component: Spinner,
+  title: "Spinner",
+};
