+++ frontend/talentsearch/src/js/components/Spinner/Spinner.tsx
