+++ frontend/talentsearch/src/js/components/Spinner/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Note: This rule seems odd, seems to think we are using
+ *       a named export here which is not the case.
+ */
+// eslint-disable-next-line no-restricted-exports
+export { default } from "./Spinner";
